Code refactoring to separate types
