+++ app/src/main/AndroidManifest.xml
@@ -44,7 +44,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="org.ocsinventoryng.android.agent.OCSAgentActivity"
+            android:name="org.ocsinventoryng.android.agent.activity.OCSAgentActivity"
             android:label="@string/app_name"
             android:process=":ocs_process">
             <intent-filter>
@@ -53,44 +53,44 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="org.ocsinventoryng.android.agent.OCSLaunchActivity"
+            android:name="org.ocsinventoryng.android.agent.activity.OCSLaunchActivity"
             android:label="@string/title_activity_ocslaunch"
             android:process=":ocs_process">
         </activity>
 
          <activity
-            android:name="org.ocsinventoryng.android.agent.OCSListActivity"
+             android:name="org.ocsinventoryng.android.agent.activity.OCSListActivity"
+             android:label="@string/title_activity_show_inventory"
+             android:process=":ocs_process">
+        </activity>
+        <activity
+            android:name="org.ocsinventoryng.android.agent.activity.OCSSectionListActivity"
             android:label="@string/title_activity_show_inventory"
             android:process=":ocs_process">
         </activity>
         <activity
-            android:name="org.ocsinventoryng.android.agent.OCSSectionListActivity"
-            android:label="@string/title_activity_show_inventory"
-            android:process=":ocs_process">
-        </activity>
-        <activity
-            android:name="org.ocsinventoryng.android.agent.OCSPrefsActivity"
+            android:name="org.ocsinventoryng.android.agent.activity.OCSPrefsActivity"
             android:label="@string/title_activity_ocsconfig"
             android:process=":ocs_process"
             android:theme="@android:style/Theme.Black">
         </activity>
 
         <service
-            android:name="org.ocsinventoryng.android.agent.OCSAgentService"
+            android:name="org.ocsinventoryng.android.agent.service.OCSAgentService"
             android:process=":ocs_process">
         </service>
         <service
-            android:name="org.ocsinventoryng.android.agent.OCSDownloadService"
+            android:name="org.ocsinventoryng.android.agent.service.OCSDownloadService"
             android:process=":ocs_process">
         </service>
 
         <receiver
-            android:name="org.ocsinventoryng.android.agent.OCSEventReceiver"
+            android:name="org.ocsinventoryng.android.agent.receiver.OCSEventReceiver"
             android:exported="true"
             android:process=":ocs_process">
         </receiver>
         <receiver
-            android:name="org.ocsinventoryng.android.agent.OCSBootReceiver"
+            android:name="org.ocsinventoryng.android.agent.receiver.OCSBootReceiver"
             android:process=":ocs_process">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED"/>
@@ -98,7 +98,7 @@
             </intent-filter>
         </receiver>
         <receiver
-            android:name="org.ocsinventoryng.android.agent.OCSInstallReceiver"
+            android:name="org.ocsinventoryng.android.agent.receiver.OCSInstallReceiver"
             android:process=":ocs_process">
             <intent-filter android:priority="10000">
                 <action android:name="android.intent.action.PACKAGE_ADDED"/>
