+++ app/src/main/java/org/ocsinventoryng/android/agent/receiver/OCSBootReceiver.java
@@ -18,7 +18,7 @@
  * along with OCSInventory-NG/AndroidAgent. if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.ocsinventoryng.android.agent;
+package org.ocsinventoryng.android.agent.receiver;
 
 import android.app.AlarmManager;
 import android.app.PendingIntent;
@@ -29,6 +29,7 @@ import android.util.Log;
 
 import org.ocsinventoryng.android.actions.OCSLog;
 import org.ocsinventoryng.android.actions.OCSSettings;
+import org.ocsinventoryng.android.agent.receiver.OCSEventReceiver;
 
 import java.util.Calendar;
