+++ app/src/main/java/org/ocsinventoryng/android/agent/service/OCSDownloadService.java
@@ -18,7 +18,7 @@
  * along with OCSInventory-NG/AndroidAgent. if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.ocsinventoryng.android.agent;
+package org.ocsinventoryng.android.agent.service;
 
 import android.app.NotificationManager;
 import android.app.PendingIntent;
@@ -36,6 +36,11 @@ import org.ocsinventoryng.android.actions.OCSProtocol;
 import org.ocsinventoryng.android.actions.OCSProtocolException;
 import org.ocsinventoryng.android.actions.OCSSettings;
 import org.ocsinventoryng.android.actions.Utils;
+import org.ocsinventoryng.android.agent.OCSDownloadIdParams;
+import org.ocsinventoryng.android.agent.OCSDownloadInfos;
+import org.ocsinventoryng.android.agent.OCSPrologReply;
+import org.ocsinventoryng.android.agent.R;
+import org.ocsinventoryng.android.agent.activity.OCSLaunchActivity;
 
 import java.io.File;
 import java.io.IOException;
